Update `mongochangestream-testing` 0.1.0 => 0.4.0
This fixes an issue that I was running into where I didn't already have a
`testing` collection in my Mongo database, and this older version of
`mongochangestream-testing` wasn't creating it and it was failing to modify the
collection to set the validation schema.

The latest version of `mongochangestream-testing` creates the collection for
you, which solved my problem.