Remove eager queue flushing
The timeout (default: 30 seconds) will ensure that the queue is eventually
flushed.

This does seem to make the tests slightly flakier, but I was able to get them
all to pass at least once. Shortening the timeout helps, so I've done that for
the resume token test case. We can do the same for other tests if we notice that
certain other ones are flaky.